+++ eslint.config.mjs
@@ -76,11 +76,11 @@ const config = createConfig([
       '**/dist/**',
       '**/docs/**',
       '**/coverage/**',
-      '.messenger-docs/**',
+      '.platform-api-docs/**',
       'merged-packages/**',
       '.yarn/**',
       'scripts/create-package/package-template/**',
-      'packages/messenger-docs/site/**',
+      'packages/platform-api-docs/site/**',
     ],
   },
   {
@@ -116,7 +116,7 @@ const config = createConfig([
       '**/tests/**/*.{js,ts}',
       'scripts/*.ts',
       'scripts/create-package/**/*.ts',
-      'packages/messenger-docs/**/*.ts',
+      'packages/platform-api-docs/**/*.ts',
     ],
     extends: [nodejs],
   },
@@ -210,7 +210,7 @@ const config = createConfig([
     },
   },
   {
-    files: ['scripts/*.ts', 'packages/messenger-docs/src/cli.ts'],
+    files: ['scripts/*.ts', 'packages/platform-api-docs/src/cli.ts'],
     rules: {
       // Scripts may be self-executable and thus have hashbangs.
       'n/hashbang': 'off',
@@ -310,7 +310,7 @@ const config = createConfig([
     },
   },
   {
-    files: ['packages/messenger-docs/**/*.{js,ts}'],
+    files: ['packages/platform-api-docs/**/*.{js,ts}'],
     rules: {
       'import-x/no-nodejs-modules': 'off',
       'n/no-process-env': 'off',