+++ packages/platform-api-docs/CHANGELOG.md
@@ -9,6 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Initial release of the messenger-docs package
+- Initial release of the platform-api-docs package
 
 [Unreleased]: https://github.com/MetaMask/core/